--- app/src/main/java/com/helio/app/ui/blind/MotionSensorSettingsFragment.java
@@ -1,32 +1,0 @@
-package com.helio.app.ui.blind;
-
-import android.os.Bundle;
-import android.text.InputType;
-import androidx.preference.EditTextPreference;
-import androidx.preference.ListPreference;
-import androidx.preference.PreferenceFragmentCompat;
-import com.helio.app.R;
-
-public class MotionSensorSettingsFragment extends PreferenceFragmentCompat {
-
-    @Override
-    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
-        setPreferencesFromResource(R.xml.motion_sensor_preferences, rootKey);
-
-        EditTextPreference namePreference = findPreference("name");
-        EditTextPreference ipPreference = findPreference("ip");
-        EditTextPreference durationSensitivityPreference = findPreference("durationSensitivity");
-        ListPreference changeIconPreference = findPreference("changeIcon");
-
-
-        assert ipPreference != null;
-        ipPreference.setOnBindEditTextListener(editText -> editText.setInputType(InputType.TYPE_CLASS_PHONE));
-
-        assert namePreference != null;
-        namePreference.setOnBindEditTextListener(editText -> editText.setInputType(InputType.TYPE_CLASS_TEXT));
-
-        assert  durationSensitivityPreference != null;
-        durationSensitivityPreference.setOnBindEditTextListener(editText -> editText.setInputType(InputType.TYPE_CLASS_NUMBER));
-
-    }
-}
