--- app/src/main/res/xml/light_sensor_preferences.xml
@@ -1,27 +1,0 @@
-<PreferenceScreen xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:andriod="http://schemas.android.com/apk/res/android">
-
-    <PreferenceCategory
-        andriod:key="light_sensor_category"
-        andriod:title="Light sensor settings">
-
-        <EditTextPreference
-            andriod:key="ip"
-            andriod:title="IP address"
-            app:useSimpleSummaryProvider="true" />
-
-        <EditTextPreference
-            andriod:key="name"
-            andriod:title="Name"
-            app:useSimpleSummaryProvider="true" />
-
-        <ListPreference
-            andriod:entries="@array/icons"
-            andriod:entryValues="@array/iconsValues"
-            andriod:key="changeIcon"
-            andriod:title="Icon"
-            app:useSimpleSummaryProvider="true" />
-
-    </PreferenceCategory>
-
-</PreferenceScreen>
